+++ library/data_structure/srq.hpp
@@ -4,11 +4,10 @@ using namespace std;
 
 // https://usaco.guide/plat/DC-SRQ
 // Not really optimized I just wrote this on a whim
-// clang-format off
 template<class T, class Q, class Op> struct RangeQuery {
     RangeQuery(const vector<T> &v, const vector<Q> &_q, Op _op, T e) :
-        ans(_q.size()), n(v.size()), d(v), lef(n, e), rig(n, e), idx(_q.size()),
-        q(_q), op(_op) {
+        ans((int)_q.size()), n((int)v.size()), d(v), lef(n, e), rig(n, e),
+        idx((int)_q.size()), q(_q), op(_op) {
         iota(begin(idx), end(idx), 0);
     }
     void solve() { _solve(0, n - 1, idx); }
@@ -48,4 +47,3 @@ template<class T, class Q, class Op>
 auto make_rq(const vector<T> &v, const vector<Q> &q, Op op, T e) {
     return RangeQuery<T, Q, Op>{v, q, op, e};
 }
-// clang-format on
