+++ library/graph/dsu_on_tree.hpp
@@ -4,7 +4,7 @@ using namespace std;
 // clang-format off
 template<class G> struct DsuOnTree {
     DsuOnTree(G &_g, int _root = 0) :
-        g(_g), n(g.size()), root(_root), sz(n, 1), euler(n), in(n),
+        g(_g), n((int)g.size()), root(_root), sz(n, 1), euler(n), in(n),
         out(n) {
         auto dfs = [&](auto &self, int v, int p) -> void {
             euler[time] = v, in[v] = time++;
